some small performance things, also moving tests
